Detach Results from GMP monolith to ResultsMixin
