+++ gvm/protocols/gmpv214/__init__.py
@@ -38,6 +38,7 @@ from gvm.protocols.gmpv208.entities.report_formats import (
     ReportFormatType,
     get_report_format_id_from_string,
 )
+from gvm.protocols.gmpv208.entities.results import ResultsMixin
 from gvm.protocols.gmpv208.entities.tasks import TaskMixin
 from gvm.protocols.gmpv208.gmpv208 import GmpV208Mixin
 from gvm.protocols.gmpv214.entities.targets import (
@@ -60,6 +61,7 @@ class Gmp(
     GmpV208Mixin,
     PortListMixin,
     ReportsMixin,
+    ResultsMixin,
     TargetMixin,
     TaskMixin,
 ):
